- move common steps to shared.rb - add passing scenarios for common misspellings - add source files to test rules
