+++ features/step_definitions/common-misspellings.rb
@@ -1,0 +1,3 @@
+def get_misspelling_filepath
+  "build/#{@dir}/"
+end
