+++ source/misspellings/index.html.md.erb
@@ -1,0 +1,10 @@
+---
+title: This page contains a misspelling of one login
+weight: 4.2
+last_reviewed_on: 2020-11-11
+review_in: 6 months
+---
+
+# If there's something strange, in your neighborhood
+
+Who you gonna call? OneLogin!
